Crate conjure_cp

Modules

ast

context

defaults

error

Top-level error types for Conjure-Oxide.

instantiate

parse

Essence parsers.

representation

rule_engine

settings

solver

A high-level API for interacting with constraints solvers.

stats

Macros

bug

Triggers a panic with a detailed bug report message, while ensuring the panic is ignored in coverage reports.

discriminant_from_name

domain_int

Creates a Domain::Int.

domain_int_ground

Creates a [GroundDomain::Int]

error

essence_expr

Parses an Essence expression into its corresponding Conjure AST at compile time.

essence_vec

Parses a sequence of Essence expressions into a vector of Conjure AST instances

into_matrix

Creates a new matrix AbstractLiteral from some Vec, optionally with some index domain.

into_matrix_expr

Creates a new matrix as an Expression from a (Rust) vector, optionally with some index domain.

matrix

Creates a new matrix AbstractLiteral optionally with some index domain.

matrix_expr

Creates a new matrix as an Expression, optionally with some index domain.

matrix_lit

Creates a new matrix Literal optionally with some index domain.

range

Creates a Range.

register_representation

throw_error

Structs

Model

An Essence model.