conjure_cp::ast::declaration

Struct DeclarationPtr

pub struct DeclarationPtr { /* private fields */ }

A shared pointer to a [Declaration].

Two declaration pointers are equal if they point to the same underlying declaration.

Id

The id of DeclarationPtr obeys the following invariants:

Declaration pointers have the same id if they point to the same underlying declaration.

The id is immutable.

Changing the declaration pointed to by the declaration pointer does not change the id. This allows declarations to be updated by replacing them with a newer version of themselves.

Ord, Hash, and Eq use id for comparisons.

Serde

Declaration pointers can be serialised using the following serializers:

DeclarationPtrFull

DeclarationPtrAsId

See their documentation for more information.

Implementations

impl DeclarationPtr

pub fn new(name: Name, kind: DeclarationKind) -> DeclarationPtr

Creates a new declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,DeclarationKind,Domain,Range};

// letting MyDomain be int(1..5)
let declaration = DeclarationPtr::new(
    Name::User("MyDomain".into()),
    DeclarationKind::DomainLetting(Domain::Int(vec![
        Range::Bounded(1,5)])));

pub fn new_var(name: Name, domain: Domain) -> DeclarationPtr

Creates a new decision variable declaration with the decision category.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,DeclarationKind,Domain,Range};

// find x: int(1..5)
let declaration = DeclarationPtr::new_var(
    Name::User("x".into()),
    Domain::Int(vec![Range::Bounded(1,5)]));

pub fn new_var_quantified(name: Name, domain: Domain) -> DeclarationPtr

Creates a new decision variable with the quantified category.

This is useful to represent a quantified / induction variable in a comprehension.

pub fn new_domain_letting(name: Name, domain: Domain) -> DeclarationPtr

Creates a new domain letting declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,DeclarationKind,Domain,Range};

// letting MyDomain be int(1..5)
let declaration = DeclarationPtr::new_domain_letting(
    Name::User("MyDomain".into()),
    Domain::Int(vec![Range::Bounded(1,5)]));

pub fn new_value_letting(name: Name, expression: Expression) -> DeclarationPtr

Creates a new value letting declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,DeclarationKind,Domain,Range, Expression,
Literal,Atom,Moo};
use conjure_cp_core::{matrix_expr,ast::Metadata};

// letting n be 10 + 10
let ten = Expression::Atomic(Metadata::new(),Atom::Literal(Literal::Int(10)));
let expression = Expression::Sum(Metadata::new(),Moo::new(matrix_expr![ten.clone(),ten]));
let declaration = DeclarationPtr::new_value_letting(
    Name::User("n".into()),
    expression);

pub fn new_given(name: Name, domain: Domain) -> DeclarationPtr

Creates a new given declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,DeclarationKind,Domain,Range};

// given n: int(1..5)
let declaration = DeclarationPtr::new_given(
    Name::User("n".into()),
    Domain::Int(vec![Range::Bounded(1,5)]));

pub fn new_record_field(entry: RecordEntry) -> DeclarationPtr

Creates a new record field declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,records::RecordEntry,Domain,Range};

// create declaration for field A in `find rec: record {A: int(0..1), B: int(0..2)}`

let field = RecordEntry {
    name: Name::User("n".into()),
    domain: Domain::Int(vec![Range::Bounded(1,5)])
};

let declaration = DeclarationPtr::new_record_field(field);

pub fn domain(&self) -> Option<Domain>

Gets the domain of the declaration, if it has one.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,Domain,Range};

// find a: int(1..5)
let declaration = DeclarationPtr::new_var(Name::User("a".into()),Domain::Int(vec![Range::Bounded(1,5)]));

assert!(declaration.domain().is_some_and(|x| x == Domain::Int(vec![Range::Bounded(1,5)])))

pub fn kind(&self) -> Ref<'_, DeclarationKind>

Gets the kind of the declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,DeclarationKind,Name,Domain,Range};

// find a: int(1..5)
let declaration = DeclarationPtr::new_var(Name::User("a".into()),Domain::Int(vec![Range::Bounded(1,5)]));
assert!(matches!(&declaration.kind() as &DeclarationKind, DeclarationKind::DecisionVariable(_)))

pub fn name(&self) -> Ref<'_, Name>

Gets the name of the declaration.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,Domain,Range};

// find a: int(1..5)
let declaration = DeclarationPtr::new_var(Name::User("a".into()),Domain::Int(vec![Range::Bounded(1,5)]));

assert_eq!(&declaration.name() as &Name, &Name::User("a".into()))

pub fn as_var(&self) -> Option<Ref<'_, DecisionVariable>>

This declaration as a decision variable, if it is one.

pub fn as_var_mut(&mut self) -> Option<RefMut<'_, DecisionVariable>>

This declaration as a mutable decision variable, if it is one.

pub fn as_domain_letting(&self) -> Option<Ref<'_, Domain>>

This declaration as a domain letting, if it is one.

pub fn as_domain_letting_mut(&mut self) -> Option<RefMut<'_, Domain>>

This declaration as a mutable domain letting, if it is one.

pub fn as_value_letting(&self) -> Option<Ref<'_, Expression>>

This declaration as a value letting, if it is one.

pub fn as_value_letting_mut(&mut self) -> Option<RefMut<'_, Expression>>

This declaration as a mutable value letting, if it is one.

pub fn replace_name(&mut self, name: Name) -> Name

Changes the name in this declaration, returning the old one.

Examples

use conjure_cp_core::ast::{DeclarationPtr, Domain, Range, Name};

// find a: int(1..5)
let mut declaration = DeclarationPtr::new_var(Name::User("a".into()),Domain::Int(vec![Range::Bounded(1,5)]));

let old_name = declaration.replace_name(Name::User("b".into()));
assert_eq!(old_name,Name::User("a".into()));
assert_eq!(&declaration.name() as &Name,&Name::User("b".into()));

pub fn detach(self) -> DeclarationPtr

Creates a new declaration pointer with the same contents as self that is not shared with anyone else.

As the resulting pointer is unshared, it will have a new id.

Examples

use conjure_cp_core::ast::{DeclarationPtr,Name,Domain,Range};

// find a: int(1..5)
let declaration = DeclarationPtr::new_var(Name::User("a".into()),Domain::Int(vec![Range::Bounded(1,5)]));

let mut declaration2 = declaration.clone();

declaration2.replace_name(Name::User("b".into()));

assert_eq!(&declaration.name() as &Name, &Name::User("b".into()));
assert_eq!(&declaration2.name() as &Name, &Name::User("b".into()));

declaration2 = declaration2.detach();

assert_eq!(&declaration2.name() as &Name, &Name::User("b".into()));

declaration2.replace_name(Name::User("c".into()));

assert_eq!(&declaration.name() as &Name, &Name::User("b".into()));
assert_eq!(&declaration2.name() as &Name, &Name::User("c".into()));