conjure_cp::ast

Module eval

Functions

bin_op

eval_constant

Simplify an expression to a constant if possible Returns: None if the expression cannot be simplified to a constant (e.g. if it contains a variable) Some(Const) if the expression can be simplified to a constant

flat_op

opt_vec_lit_op

opt_vec_op

tern_op

un_op

unwrap_expr

vec_lit_op

vec_op