conjure_cp::ast::matrix

Trait MatrixValue

pub trait MatrixValue:
    Sized
    + AbstractLiteralValue
    + From<AbstractLiteral<Self>>
    + Biplate<AbstractLiteral<Self>> {
    // Required methods
    fn as_nested_matrix(&self) -> Option<&AbstractLiteral<Self>>;
    fn into_nested_matrix(self) -> Result<AbstractLiteral<Self>, Self>;
}

Things that can appear inside a matrix.

This is a helper trait to unify matrix operations on Expression::AbstractLiteral and AbstractLiteral<Literal>

Required Methods

fn as_nested_matrix(&self) -> Option<&AbstractLiteral<Self>>

If this element is a nested matrix, return a reference to it

fn into_nested_matrix(self) -> Result<AbstractLiteral<Self>, Self>

If this element is a nested matrix, consume it and return the matrix

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".