conjure_cp::ast

Struct SubModel

pub struct SubModel { /* private fields */ }

A sub-model, representing a lexical scope in the model.

Each sub-model contains a symbol table representing its scope, as well as a expression tree.

The expression tree is formed of a root node of type Expression::Root, which contains a vector of top-level constraints.

Implementations

impl SubModel

pub fn new(parent: Rc<RefCell<SymbolTable>>) -> SubModel

Creates a new [Submodel] as a child scope of parent.

parent should be the symbol table of the containing scope of this sub-model.

pub fn symbols_ptr_unchecked(&self) -> &Rc<RefCell<SymbolTable>>

The symbol table for this sub-model as a pointer.

The caller should only mutate the returned symbol table if this method was called on a mutable model.

pub fn symbols_ptr_unchecked_mut(&mut self) -> &mut Rc<RefCell<SymbolTable>>

The symbol table for this sub-model as a mutable pointer.

The caller should only mutate the returned symbol table if this method was called on a mutable model.

pub fn symbols(&self) -> Ref<'_, SymbolTable>

The symbol table for this sub-model as a reference.

pub fn symbols_mut(&mut self) -> RefMut<'_, SymbolTable>

The symbol table for this sub-model as a mutable reference.

pub fn root(&self) -> &Expression

The root node of this sub-model.

The root node is an Expression::Root containing a vector of the top level constraints in this sub-model.

pub fn root_mut_unchecked(&mut self) -> &mut Expression

The root node of this sub-model, as a mutable reference.

The caller is responsible for ensuring that the root node remains an Expression::Root.

pub fn replace_root(&mut self, new_root: Expression) -> Expression

Replaces the root node with new_root, returning the old root node.

Panics

If new_root is not an Expression::Root.

pub fn constraints(&self) -> &Vec<Expression>

The top-level constraints in this sub-model.

pub fn clauses(&self) -> &Vec<CnfClause>

The cnf clauses in this sub-model.

pub fn constraints_mut(&mut self) -> &mut Vec<Expression>

The top-level constraints in this sub-model as a mutable vector.

pub fn clauses_mut(&mut self) -> &mut Vec<CnfClause>

The cnf clauses in this sub-model as a mutable vector.

pub fn replace_constraints( &mut self, new_constraints: Vec<Expression>, ) -> Vec<Expression>

Replaces the top-level constraints with new_constraints, returning the old ones.

pub fn replace_clauses(&mut self, new_clauses: Vec<CnfClause>) -> Vec<CnfClause>

Replaces the cnf clauses with new_clauses, returning the old ones.

pub fn add_constraint(&mut self, constraint: Expression)

Adds a top-level constraint.

pub fn add_clause(&mut self, clause: CnfClause)

Adds a cnf clause.

pub fn add_constraints(&mut self, constraints: Vec<Expression>)

Adds top-level constraints.

pub fn add_clauses(&mut self, clauses: Vec<CnfClause>)

Adds cnf clauses.

pub fn add_symbol(&mut self, decl: DeclarationPtr) -> Option<()>

Adds a new symbol to the symbol table (Wrapper over SymbolTable.insert)

pub fn into_single_expression(self) -> Expression

Converts the constraints in this submodel to a single expression suitable for use inside another expression tree.

If this submodel has no constraints, true is returned.

If this submodel has a single constraint, that constraint is returned.

If this submodel has multiple constraints, they are returned as an and constraint.