conjure_cp::ast

Struct SymbolTable

pub struct SymbolTable { /* private fields */ }

The global symbol table, mapping names to their definitions.

Names in the symbol table are unique, including between different types of object stored in the symbol table. For example, you cannot have a letting and decision variable with the same name.

Symbol Kinds

The symbol table tracks the following types of symbol:

Decision Variables

find NAME: DOMAIN

See DecisionVariable.

Lettings

Lettings define constants, of which there are two types:

Constant values: letting val be A, where A is an Expression.

A can be any integer, boolean, or matrix expression. A can include references to other lettings, model parameters, and, unlike Savile Row, decision variables.

Constant domains: letting Domain be domain D, where D is a Domain.

D can include references to other lettings and model parameters, and, unlike Savile Row, decision variables.

Unless otherwise stated, these follow the semantics specified in section 2.2.2 of the Savile Row manual (version 1.9.1 at time of writing).

Implementations

impl SymbolTable

pub fn new() -> SymbolTable

Creates an empty symbol table.

pub fn with_parent(parent: Rc<RefCell<SymbolTable>>) -> SymbolTable

Creates an empty symbol table with the given parent.

pub fn lookup_local(&self, name: &Name) -> Option<DeclarationPtr>

Looks up the declaration with the given name in the current scope only.

Returns None if there is no declaration with that name in the current scope.

pub fn lookup(&self, name: &Name) -> Option<DeclarationPtr>

Looks up the declaration with the given name, checking all enclosing scopes.

Returns None if there is no declaration with that name in scope.

pub fn insert(&mut self, declaration: DeclarationPtr) -> Option<()>

Inserts a declaration into the symbol table.

Returns None if there is already a symbol with this name in the local scope.

pub fn update_insert(&mut self, declaration: DeclarationPtr)

Updates or adds a declaration in the immediate local scope.

pub fn return_type(&self, name: &Name) -> Option<ReturnType>

Looks up the return type for name if it has one and is in scope.

pub fn return_type_local(&self, name: &Name) -> Option<ReturnType>

Looks up the return type for name if has one and is in the local scope.

pub fn domain(&self, name: &Name) -> Option<Domain>

Looks up the domain of name if it has one and is in scope.

This method can return domain references: if a ground domain is always required, use SymbolTable::resolve_domain.

pub fn resolve_domain(&self, name: &Name) -> Option<Domain>

Looks up the domain of name, resolving domain references to ground domains.

See SymbolTable::domain.

pub fn into_iter_local(self) -> LocalIntoIter

Iterates over entries in the local symbol table only.

pub fn extend(&mut self, other: SymbolTable)

Extends the symbol table with the given symbol table, updating the gensym counter if necessary.

pub fn gensym(&mut self, domain: &Domain) -> DeclarationPtr

Creates a new variable in this symbol table with a unique name, and returns its declaration.

pub fn parent_mut_unchecked(&mut self) -> &mut Option<Rc<RefCell<SymbolTable>>>

Gets the parent of this symbol table as a mutable reference.

This function provides no sanity checks.

pub fn get_representation( &self, name: &Name, representation: &[&str], ) -> Option<Vec<Box<dyn Representation>>>

Gets the representation representation for name.

Returns

None if name does not exist, is not a decision variable, or does not have that representation.

pub fn representations_for( &self, name: &Name, ) -> Option<Vec<Vec<Box<dyn Representation>>>>

Gets all initialised representations for name.

Returns

None if name does not exist, or is not a decision variable.

pub fn get_or_add_representation( &mut self, name: &Name, representation: &[&str], ) -> Option<Vec<Box<dyn Representation>>>

Gets the representation representation for name, creating it if it does not exist.

If the representation does not exist, this method initialises the representation in this symbol table, adding the representation to name, and the declarations for the represented variables to the symbol table.

Usage

Representations for variable references should be selected and created by the select_representation rule. Therefore, this method should not be used in other rules. Consider using get_representation instead.

Returns

None if name does not exist, is not a decision variable, or cannot be given that representation.