conjure_cp::parse

Crate tree_sitter

Parse Essence using the tree-sitter-essence tree-sitter grammar.

Modules

diagnostics

errors

parser

parser_legacy

Macros

child

Get the i-th child of a node, or return a syntax error with a message if it doesn’t exist.

field

Get the named field of a node, or return a syntax error with a message if it doesn’t exist.

named_child

Get the i-th named child of a node, or return a syntax error with a message if it doesn’t exist.

Enums

EssenceParseError

Functions

parse_essence

parse_essence_file

parse_essence_file_native

Parse an Essence file into a Model using the tree-sitter parser.

parse_essence_with_context

parse_expr

parse_exprs

parse_literal