conjure_cp

Module representation

Modules

default_impls

Default implementations of some Repr trait methods using functors; These methods are used by the register_repr! macro

errors

registry

stored

types

util

Macros

register_representation

Structs

ReprStore

Enums

ReferenceReprError

Either ReprInitError | ReprInstantiateError | ReprSelectError

ReprDownError

ReprError

Either ReprInitError or ReprInstantiateError

ReprInitError

Errors that can be thrown by ReprDomainLevel::init

ReprInstantiateError

Errors that can be thrown by ReprDomainLevel::instantiate

ReprSelectError

ReprUpError

Traits

ReprAssignment

ReprDeclLevel

ReprDomainLevel

ReprRule

ReprRuleStored

ReprStateStored

Representation

Encodes the two-way relationship between a non-atomic variable and multiple auxiliary variables.

Functions

get_applicable_repr_by_short_name

Look up a representation by short name that is applicable to decl.

get_repr_by_name

get_repr_by_short_name

Look up a representation rule by its Essence short name (e.g. "packed", "occurrence").

get_repr_rule

Gets the representation rule named name.

get_repr_rules

Type Aliases

ReprResult

ReprRulePtr