conjure_cp::representation

Module default_impls

Default implementations of some Repr trait methods using functors; These methods are used by the register_repr! macro

Functions

domain_size

Returns the number of values in a representation-variable domain. Domains whose size is unknown or does not fit in usize rank as least compact.

instantiate_default_impl

Implement ReprDomainLevel::instantiate as a functor S<DomainPtr> -> S<DeclarationPtr>. Needs the following additional arguments:

lookup_via_default_impl

Implement [ReprDeclLevel::lookup_via] as a functor S<DeclarationPtr> -> S<Literal>.

to_domain_level_default_impl

Implement [ReprDeclLevel::to_domain_level] as a functor S<DeclarationPtr> -> S<DomainPtr>.