conjure_cp::rule_engine

Function get_all_rule_sets

pub fn get_all_rule_sets() -> Vec<&'static RuleSet<'static>>

Get all rule sets Returns a Vec of static references to all rule sets registered with the register_rule_set macro. Rule sets are not guaranteed to be in any particular order.

Example

use conjure_cp_core::rule_engine::register_rule_set;
use conjure_cp_core::rule_engine::get_all_rule_sets;

register_rule_set!("MyRuleSet", ("AnotherRuleSet"));
register_rule_set!("AnotherRuleSet", ());

println!("Rule sets: {:?}", get_all_rule_sets());

This will print (if no other rule sets are registered):

Rule sets: [
  RuleSet { name: "MyRuleSet", rules: OnceLock { state: Uninitialized }, dependencies: ["AnotherRuleSet"] },
  RuleSet { name: "AnotherRuleSet", rules: OnceLock { state: Uninitialized }, dependencies: [] }
]