conjure_cp::ast

Struct Moo

pub struct Moo<T> { /* private fields */ }

A clone-on-write, reference counted pointer to an AST type.

Cloning values of this type will not clone the underlying value until it is modified, e.g., with Moo::make_mut.

Unlike Rc and Arc, trait implementations on this type do not need to preserve shared ownership - that is, two pointers that used to point to the same value may not do so after calling a trait method on them. In particular, calling Uniplate methods may cause a clone-on-write to occur.

Note: like BoxandRc, methods on Mooare all associated functions, which means you have to call them as, e.g.Moo::make_mut(&value)instead ofvalue.make_mut(). This is so that there are no conflicts with the inner type T`, which this type dereferences to.

Implementations

impl Moo<Domain>

pub fn resolve(&self) -> Option<Moo<GroundDomain>>

pub fn union(&self, other: &Moo<Domain>) -> Result<Moo<Domain>, DomainOpError>

Convenience method to take Domain::union of the Domains behind two DomainPtrs and wrap the result in a new DomainPtr.

pub fn intersect( &self, other: &Moo<Domain>, ) -> Result<Moo<Domain>, DomainOpError>

Convenience method to take Domain::intersect of the Domains behind two DomainPtrs and wrap the result in a new DomainPtr.

impl<T> Moo<T>

pub fn new(value: T) -> Moo<T>

Constructs a new Moo<T>.

impl<T> Moo<T>
where T: Clone,

pub fn make_mut(this: &mut Moo<T>) -> &mut T

Makes a mutable reference into the given Moo.

If there are other Moo pointers to the same allocation, then make_mut will clone the inner value to a new allocation to ensure unique ownership. This is also referred to as clone-on-write.

pub fn unwrap_or_clone(this: Moo<T>) -> T

If we have the only reference to T then unwrap it. Otherwise, clone T and return the clone.

Assuming moo_t is of type Moo<T>, this function is functionally equivalent to (*moo_t).clone(), but will avoid cloning the inner value where possible.